+++ frontend/features/profile/components/AccountSetting.tsx
@@ -25,7 +25,7 @@ import * as z from 'zod'
 import { ContentWrapperLoading } from '@/features/shared/ContentWrapperLoading'
 import { useUpdateProfile } from '../hooks/use-update-profile'
 import { SubmitButton } from '@/features/shared/SubmitButton'
-import RenderIf from '@/components/RenderIf'
+import RenderIf from '@/features/shared/RenderIf'
 import UploadAvatarModal from '@/features/profile/components/modal/UploadAvatarModal'
 import { toast } from 'react-toastify'
 import { useTheme } from 'next-themes'
